+++ rumqttd/src/lib.rs
@@ -55,6 +55,7 @@ pub type AuthHandler = Arc<
 pub struct Config {
     pub id: usize,
     pub router: RouterConfig,
+    pub tcp: Option<HashMap<String, ServerSettings>>,
     pub v4: Option<HashMap<String, ServerSettings>>,
     pub v5: Option<HashMap<String, ServerSettings>>,
     pub ws: Option<HashMap<String, ServerSettings>>,
@@ -116,6 +117,8 @@ pub struct ServerSettings {
     pub tls: Option<TlsConfig>,
     pub next_connection_delay_ms: u64,
     pub connections: ConnectionSettings,
+    #[serde(default)]
+    pub supported_protocol: SupportedProtocol,
 }
 
 impl ServerSettings {
@@ -253,4 +256,15 @@ pub enum MetricType {
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct MetricSettings {
     push_interval: u64,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, Default)]
+pub enum SupportedProtocol {
+    #[serde(rename = "all")]
+    #[default]
+    All,
+    #[serde(rename = "v3.1.1")]
+    V4,
+    #[serde(rename = "v5")]
+    V5,
 }